refactor: created StorymapMetaEdit for the common information of storymap to initialize
